Stub candidate access token request
If a candidate already exists in Dynamics365 we want to be able to pull out the
candidate details in a secure way. To achieve this the client will send the
candidate email address, in addition to 2-3 extra pieces of personal
information. If these match an existing candidate then we will email an access
token (PIN code) that can be used to retrieve the full candidate details.

This commit adds the controller action stub for this function, in addition to
the `CandidateAccessTokenRequest` model and associated validation (using
FluentValidation).

Also introduces Swashbuckle so that we can generate the Swagger documentation
from the source code going forward as well as common testing packages (Moq and
FluentAssertions).