Stub mailing list subscription request
The Get into Teaching website will have a mailing list that users are able to
sign up to; when this happens we will flag the consent in Dynamics365.

This commit stubs out a controller action for this function as well as
introducing a `CandidateIdentification` model and associated validations. The
model encapsulates the information we require for identification purposes when
subscribing a user to a mailing list, or adding them as an attendee to an event
(this will come later).