Stub type requests
A candidate has several attributes that will be presented in forms as select
controls containing many possible values. Instead of hard-coding these values
in the client, we are exposing `/api/types` endpoints that can return the
available options. Initially to support the Get into Teaching TTA service we
will need `countries` and `teaching subjects`. This commit stubs out the
controller and actions for these types.