Stub TTA candidate rquests
The Teacher Training Adviser (TTA) service will enable candidates to sign up so
that they can speak to an adviser. The information we need about a candidate in
order to sign up to this service will be different from other services (such as
Schools Experience). This commit stubs out a controller with TTA-specific
candidate actions.

A client will be able to request an existing TTA candidate with a valid access
token (PIN code) or upsert a candidate (which will register them for the TTA
service).