Add service to generate candidate access tokens
We need to be able to generate access tokens that can be exchanged for existing
candidate information. This commit introduces a token generation service based
on the Timed One Time Password (TOTP) RFC 6238.

A new token is generated ever 30 second 'step' (as per the RFC) and the service
is currently configured so that tokens remain valid for a step windows size of
2; this means that the current step, plus 2 steps (1 minute) prior and 2 steps
future will be valid. Essentially, it results in the candidate having roughtly
1 minute (plus at most another 30 seconds) to receive the email and enter the
code. We can customise this in the future by increasing the window size to
allow more time (the step size should remain at 30 seconds as per the RFD).

An environment variable `TOTP_SECRET_KEY' must be set; this is used for the
TOTP generation. We use the candidate email address in addition to the
`TOTP_SECRET_KEY` as the TOTP secret to ensure that candidates with different
emails get different tokens generated when inside the same time step.