Verify access token on existing candidate request
When a client requests information about an existing candidate they will send a
TOTP access token. This commit adds the verification logic to the
`TeacherTrainingAdvisor.CandidatesController` to ensure the access token is
verified before proceeding to return candidate information.

The candidate email address has been added to the request header since TOTPs
are not unique per-request and we must use the candidate email address as part
of the TOTP secret in order to get different codes for different candidates
within the same time step.

If we didn't do this, then we would be vulnerable to a candidate generating a
token for their own email address _and_ another candidate's email address --
the token would be valid/the same for both. As per the RFC the HMAC is
truncated, which means that two different emails _could_ result in the same
token, but chances of this should be very slim (and the tokens themselves are
only 'live' for a handful of minutes which means it should be even slimmer).