Include Notify package and wrap in an adapter
We will be using the GOV.UK Notify service in order to send PIN code emails out
to candidates. There is a C# client library that we are using, however it
doesn't provide an interface and the methods are not virtual, which means we
won't be able to mock them. This commit adds a simple adapter around the client
so that we can mock out the interface in later tests.