Adds a basic notification service
We are adding a basic notification service that configures and calls out to the
GOV.UK Notify client. As a first pass we are opting to not push this to a
background process and instead are calling the async function to send the email
without blocking. We will catch if the email fails to send in another thread
and log out the error. Going forward we will monitor this and if the
reliability of the email service warrants it, we can move the logic to a
background process with retry capabilities.