Add a CrmService for interacting with Dynamics365
Adds a service that encapsulates the interactions with the Dynamics365
instance. Currently it supports getting countries and teaching subjects,
however it will be extended to write back to Dynamics as well.

Use the `CdsServiceClient` SDK which is currently in ALPHA. We may swap over to
.NET Framework in the future so that we can use the stable `CrmServiceClient`
instead.

Wrap the `IOrganizationServiceContext` in an adapter so that we can mock it in
unit tests and execute the LINQ queries against an in-memory `IQueryable`.

AutoMapper is used to translate between `Entity` objects returned from the SDK
and our models (in this case the `TypeEntity` model.