Add GetCandidate method to CrmService
We need to be able to retrieve a candidate from the CRM by email address so
that we can match it to an incoming `CandidateAccessTokenRequest`. There may be
multiple candidates with the same email address, so we will take the latest.

Add basic `Candidate` model with minimal attributes (to be fleshed out in a
later PR where we return the `Candidate` from our API).

Add profile to map from Dynamics `Entity` to our `Candidate` model.