Remove async wrapper around CRM interactions
The async wrapper around the CRM integration is just adding unnecessary
complexity at the moment. If we find that the service could benefit from it we
can add it back in at the end; it would be cleaner to do so at the controller
level to avoid the use of `async`/`await` being peppered throughout the rest of
the application.