Update TypeEntity id to be dynamic
We are using `TypeEntity` to map from a 'lookup' (`Entity`) object as well as a
'pick list' (`PickListItem`) object. Both of these objects end up as options in
a select element of a form and they can be represented with an identifier and
value. By making the `Id` attribute of `TypeEntity` dynamic it will accept both
a `Guid` and `int` (which are used by lookup and pick list objects,
respectively).

Also updates the auto mapper profile to support mapping from a `PickListItem`
as well as the existing `Entity`.