Make EntityType operations on CRM generic
We are going to have a lot of lookups/pick list items queried via the CRM going
forward. Instead of maintaining a separate method for each one this commit
refactors into two generic methods (one for lookup items and one for pick list
items). This does result in the entity name and attribute names leaking out
into the `TypesController`, however we can look at encapsulating those in
another class as part of a future PR.