Perform candidate matching in CrmService
Candidate matching is currently performed in the controller, however we are
going to be matching requests to existing candidates in other places so it
makes sense to perform that operation within the `CrmService` so it can be
reused.

Update candidate matching to up to 20 candidates that match the email,
returning the latest candidate that matches successfully.