Update CrmService to retrieve associated Candidate entities
As part of the Get into Teaching registration form we will be asking candidates
about their past teaching positions and qualifications. The information is
stored as separate entities in Dynamics - this commit updates the `CrmService`
to query the entities when retrieving a candidate (running them through the
AutoMapper profile to get back an equivilent API model).