Only accept a privacy policy once
Update mapping operation to enable subclass to hook into mapping and decide if
a property should be mapped or not.

Method added to `CrmService` to check if a candidate has already accepted a
privacy policy - called in the `Candidate` model to skip mapping a
`CandidatePrivacyPolicy` entity if they have already accepted this policy.

Encapsulate `IOrganizationServiceAdapter` in the `CrmService` - previously this
was being exposed to enable some of the mapping functionality in `BaseModel`,
however it's cleaner if we proxy the required methods via `CrmService` and then
encapsulate the adapter.

General bits of tidy up after refactoring work in #22.