Add in-memory cache for CrmService
The in-memory cache will store responses from the CRM that are used in
validation logic, so that we can still accept requests from a client when the
CRM is offline. This covers:

- Lookup items
- PickList items
- Privacy policy entities

The cache will be refreshed every 3 hours, however if the CRM is offline when a
refresh is attempted the stale data will remain available in the cache so that
the API can continue to function.

As this cache is in-memory, if the API server restarts when the CRM is offline
it will not be able to function until the CRM comes back online. We expect the
chance of this happening to be slim and will monitor our logs to see if it is
worth the effort of implementing a persistent cache.