Add PostcodeService
We need to be able to validate a postcode coming from a teaching event search
request against the list of known postcodes that we can geocode.

A user can also filter by events within a certain radius of a postcode, so
there is a method to calculate the distance between two postcodes.

Postcode data is stored in-memory; this causes the application to be slow when
booting, so we will likely move this to a database in a future PR.