WIP: Create candidates via Web API
Work in progress: candidates are created, however the qualifications are
not correctly assigned to the candidate (past teaching positions work
fine, strangely).

Also, the logic to ensure a candidate only accepts a privacy policy once
has not yet been implemented (find TODO:'s for what's left).

As the Simple OData Client sends null GUID values on insert operations
we have to strip these out - this is done in a pre-request hook on the
client settings.

The other concession with this method over the SDK is that the Simple
OData Client does not support deep insert, so we have to make multiple
requests to save an object. The batch API doesn't work correctly with
Dynamics either - it fails to link the objects on the segment IDs.