Move transactions to background jobs
We currently send data to dynamics within the request/response cycle.
This is fine if Dynamics is online, however it can be offline for
maintenance for reasonably long periods of time. To ensure the
transactions complete this commit adds Hangfire to queue the transaction
requests and retry them in the event of failure.

The jobs will be retried once per hour, a maximum of 24 times. If the
job is not successful it will be removed from storage and we will
attempt to email the candidate informing them that the transaction has
failed.

Hangfire is configured to use in-memory storage for development (and
also runs a shorter retry/give up cycle). In production we are using
PostgreSQL as this is supported by the free version of Hangfire (Redis
requires a Hangfire susbcription).

The entire candidate/event registration object is serialized as part of
the job for simplicity - this should be fine given the size of those
models just now, but in the future we may want to store them separately
in Postgres and only pass the ID of the object to the job.

PerformContextAdapter enables the mocking and unit testing of the
retry/give up behavior.