Migrate postcode data into Postgres
There are roughtly 1.7 million UK postcodes that we currently read from
a CSV file and store in memory to do geolocation comparisons when
searching for events by radius.

This commit introduces Entity Framework, backed by Postgres and a
`Location` model; the postcode and associated geolocation are read from
the CSV and seeded into the database (if not yet present) when the
application starts.

The `LocationService` has been updated to query postcode information
from the database and the in-memory store has been removed.

Entity Framework is configured to use an in-memory SQLite database when
in development.

Going forward it will make sense to also put event information into
Postgres so that we can run a query to find events within a geographical
region, instead of querying postcode information individually for each
event as we currently do.