Build connection strings from VCAP_SERVICES env
We previously had the connection strings hard-coded thinking that the
app could connect using the Postgres service identifier, but this
doesn't work how I was expecting.

Instead, this commit uses the `VCAP_SERVICES` env variable that will be
available to the application; we build a connection string dynamically
from this.