Sync teaching events to Postgres
Teaching events are currently cached in-memory after the initial request
to get events via the API and periodically refreshed on a fixed
interval. The issues with this mechanism are that there are a lot of
events, so it uses a bunch of memory, and the first request or
subsequent requests that refresh the cache will be slow as this is done
in the request/reponse cycle.

This commit decouples event reading from updating our cache of the CRM
events. Instead, we periodically sync the events from the CRM into
Postgres on a background job (daily). That way any requests for events
only hit our Postgres database and should return quickly.

A follow up commit will refactor the event querying logic to be
performed in the database (currently events are loaded into memory
before filtering)..