Ensure BaseModel subtypes have required constructor
As part of mapping Dynamics entities to our models we need to
dynamically instantiate BaseModel subtypes using reflection; if a
subtype doesn't have the constructor we expect a runtime error will
occur. As it's used with reflection we don't get a compile time error if
the constructor is missing, so this commit introduces a test that will
ensure the constructor is present.