Enable new candidates to register for events
Previously we only checked for existing candidates when registering a
candidate for an event. Instead, we attempt to get the matching,
existing candidate if they exist and if not then a new candidate entry
is created and registered with the event.

If the registration job fails after creating a candidate then on the
next run the candidate will be matched, so we won't end up creating a
bunch of duplicate candidates in the event of job failures.