Filter teaching events at the database layer
Currently we load all teaching events into memory and then apply the
filtering from the search request. This is inefficient; instead, this
commit stores the lat/lng coordinates against each teaching event and
performs the geospatial distance calculation at the database layer.

We do an approximate distance calculation to avoid the complexity of
projecting coordinates from a planar coordinate system to correctly
compare with the configured SRID. It should be accurate enough for our
purposes (as we're only using postcode granualrity anyway).

Also removes `LocationService`, consolidating it into a new `Store` that
will be responsible for abstracting database interactions.