Wrap environment vars in an adapter
We can't mock the environment variables on `System.Environment` and changing
them during the test run causes flakiness. This commit adds an adapter `Env` to
abstract the environment variables and allow mocking. The environment is still
changed in a couple of places, but this should resolve much of the flakiness,
which isn't too bad in general anyway.