Add health check endpoint
The health of the system can be checked by making a GET request to
/api/operations/health_check - it will respond with a JSON payload:

```
{
  "database": true,
  "crm": true,
  "notify": true,
  "status": "healthy"
}
```

If `crm` or `notify` are `false` then the status is `degraded` (as we are
operating off potentially stale CRM data and aren't able to send candidate PIN
codes - we have no control over the availability of these services, though). If
`database` is `false` then the service status will be `unhealthy`.