Prevent flakeyness in tests changing environment
Most of the test cases are able to mock `IEnv` to change the environment,
however a few that are static methods/classes still call
`Environment.GetEnvironmentVariable` directly. If a test changes the
environment with `Environment.SetEnvironmentVariable` it can cause conflicts
with other tests that rely on the same env var. This introduces a collection to
disable parallelization of theses tests.