Add endpoint to generate mapping information
We want to be able to provide the CRM team with an up-to-date list of fields
that we map to in Dynamics 365 for each of our API models. This commit adds an
endpoint to generate this mapping information using the existing field
attributes that we decorate the API models with to map to/from Dynamics
entities.