Run location sync periodically
We currently seed the locations table with postcode geolocation data from a
static CSV file vendored in the source code. Postcodes will change over time,
so this commit moves the location seeding to a background job that fetches the
latest CSV, unzips it and adds any new postcodes into our database on a weekly
basis.