Register CdsServiceClient as transient/cloned
The CdsServiceClient is not asynchronous and there is overhead in creating a
new instance (as it will go away and authenticate each time and download
metadata). To ensure that we can handle multiple asyc requests hitting the CRM
this commit registers a singleton CdsServiceClient that we then clone as a
transient dependency.