Swap authorization to authentication
We are currently using policy-based authorization to determine if a request can
access the API (if it has the correct shared secret). ASP.NET Core assumes you
set an authentication provider if you are using authorization (it works if you
don't, but you don't get the correct 401 unauthorized response on failures, for
some reason). Instead, this commit drops the authorization handler in favour of
an authentication handler, which ends up being simpler anyway.