Run sync operations in series
Previously I updated the sync methods to run in parallel for better
performance, however this instroduced an issue as the DbContext is shared
across each method in the `Store` (multiple threads accessing the same context
is not supported).

Change back to running the sync operations in series; it only takes about 10
seconds to do a full sync so performance isn't an issue.