Only expose Hangfire to Prometheus on one instance
Each application instance runs a number of Hangfire workers against the same
shared job storage. If we expose the Hangfire statistics to Prometheus for
scraping on each instance it will aggregate the results, which would be
incorrect. This commit uses the `CF_INSTANCE_INDEX` environment variable to
ensure only one instance is exposing the Hangfire job stats.