Clean up use of Env throughout
At the minute we use a mixture of the dependency injected `IEnv` and creating
new instances of `Env` where it's harder to inject one to. This commit exposes
`IEnv` as a dependency throughout so that we are using it consistently. This
also means we don't need to mock `Environment.GetEnvironmentVariable` so the
thread safety collection can be reduced to just the `EnvTests`.