Set fixed number of Hangfire workers
The Postgres connection limit (100) is being hit in the test environment
shortly after Hangfire starts up. By default the number of workers is a
multiple of the number of cores the instance has - we don't know what this is,
but assuming 20 cores it would result in `20 * 5 = 100` connections being
created and the limit being hit. This commit fixes the number of workers to 10,
which should be sufficient for the application.