Add endpoint to retrieve callback booking quotas
Retrieve the callback booking quotas from Dynamics. Returns the booking quotas
for the next 14 days, in alignment with the existing TTA process (matches the
Azure function the CRM team provided). They also allow filtering to a specific
number of weekdays, but I'm not sure if this is a requirement of the new
process so I've left that out for now.