Update event registration to use manual matchback
Instead of matching the candidate automatically on event registration we want
to manually make the user verify their email address first. The event
registration endpoint now accepts a `candidateId` that can be used to add an
existing candidate to an event (the id should be obtained from a matchback
request and only used once a user has verified the email).

Also adds additional fields to the candidate entity to match the updated event
registration process.