Avoid dependency on CrmService for integration test
The integration test previous tested the `OperationsController`, however that
now has a dependency on the `CrmService` (which throws an exception since we're
not configuring it in the test environment). Instead, this commit changes the
integration test to request the latest privacy policy - the store is empty so
it simply asserts a 200 response, but it's also checking the authentication
layer now since the `PrivacyPoliciesController` is a protected route.