Support ETags for content synced from CRM
Content is currently synced from the CRM on a weekly basis; any requests made
should be cachable up until the moment a new sync has been performed. To
achieve this, the commit generates an `ETag` based on the next CRM sync job
execution date (which changes after a successful sync) along with the request
path. When an `If-Not-Modified` header is sent containing the current `ETag` we
immediately respond with a 304 (not modified) header, avoiding the overhead of
calling the action and querying the data. If an incorrect/outdated `ETag` is
present (or no `If-Not-Modified` header is provided) then the request completes
as normal and returns an `ETag`.

Any non-200 responses are stripped of their `ETag` in the `OnActionExecuted`
callback to ensure bad responses are never cached.