Generate v3 OpenAPI docs in development
The UI created by Swashbuckle for V2 Swagger docs is buggy (if you paste in
JSON as a request body it doesn't always let you submit the form, but it
doesn't give you an error either). This commit updates the Swashbuckle config
to generate the more reliable V3 UI when in development, but creates V2 in
other environments as this is needed to use the Swagger codegen tool for Ruby.