Update PhoneCall entity to match schema changes
Certain fields of the `PhoneCall` model are populated using the data already
collected in the `Candidate` model. This commit hooks into the mapping process
to set the values on the `PhoneCall` before mapping it to an `Entity`.

Add a number of hidden fields with default values that the CRM expects.