Remove flakey integration test
This is being flakey because of the background jobs that start to run - if the
request finishes before the job then it passes, if not the job errors out
because Hangfire can't connect to anything and the test fails. We are looking
at introducing more thorough end-to-end tests down the line anyway, so removing
this for now.