Add TeachingEventAddAttendeeRequest model
Encapsulate all the information required to register for a teaching event.
There are still some questions around the consent settings for this flow, but
they match the spreadsheet the CRM team have provided as it currently stands.
Also allows opting in to the events and marketing emails. We still need to
handle if the candidate already has a subscription to these services.