Prevent exception when running migrations
The GetIntoTeachingDbFactory class requires an empty constructor and it also
runs the `DbConfiguration` -- even though it doesn't connect to a database it
does need to be able to create the connection string successfully. It was
previously throwing an exception as the `VCAP_SERVICES` env variable was not
available; this commit udpates the factory to mock out the `VCAP_SERVICES` env
var.