Document validation error mapping in request models
An error may be present on the request model or a hidden `BaseModel` model that
it maps to. The client needs to check for the incluision of the request model
field in any error keys when reconciling errors with particular fields. This
commit documents the behaviour in the endpoints that use request models.