Retrieve TeachingEvent by ReadableId instead of Id
Currently we put the event id, which is a long Guid, into the URL in the
front-end to be able to retrieve the specific event from the API when the page
loads. We want more SEO friendly URLs, so we're changing the API to expose the
readable id of the event (which is a CRM-created permalink) and then using this
as the lookup in the `GetTeachingEvent` operation.

An example of the readable id: `439_-_Online_TTT_-_London-Summer_20203437949823`