Return AuthenticateResult.NoResult() if no Authorization header present
The auth schema is called for all requests, regardless of if they require
authentication. If a request does not contain an `Authorization` header we
returned a failure result, which causes a log message to be written out (twice,
as we were logging it independently as well). This is not actually an issue for
endpoints that do not require authentication - instead, we should be returning
`AuthenticationResult.NoResult()`.