Accept any valid postcode on event search
Previously we validated the postcode on event search by checking to ensure it's
in our local database of locations (geocodable postcodes). As our local data
set may not be completely up to date, and we now fallback to Google APIs, it's
possible that a valid postcode is passed but not in our database. To account
for this, we validate the postcode format (instead of if it exists) and if we
fail to geocode the postcode later on in the request cycle we return an empty
data set.