Add additional properties to TeachingEventRegistration
According to the latest schema we should be passing `msevtmgt_iscanceled` and
`msevtmgt_registrationnotificationseen` as `false` to new teaching event
registrations. There is also two new fields `msevtmgt_publishingstate` and
`msevtmgt_syncedwithprovider`, however after consulting with the CRM team we
should be fine to leave these as the defaults (which are `No` and `Pending`).