Exclude draft/past events from sync operation
We do not want to display events on the website if they are not in an
open/closed status (so draft or other). Events that are in the past should also
be excluded from displaying; we are checking the event end date to ensure it is
in the future, but the criteria around this may change (it may transpire that
we want to display events until the day after they end).