Only de-duplicate on active subscriptions
If a Candidate already has a subscription of the same type that is active, then
we shouldn't create another instance of the subscription. Previously, we were
not taking into account the `statecode` of the Candidate's existing
subscriptions and therefore a new Subscription was not being created when it
should have been (if a user subscribed, opted-out and then re-subscribed).