Exclude early years/other event types
It has been decided that we do not want to show Early Year/Other event types in
the Get into Teaching website, so this commit removes them.

Explicitly assert event types in the test to ensure accidently removing an enum
from `TeachingEvent` doesn't result in the events of that type from being
excluded (the connection was not explicit previously).