Move de-duplication logic into TeachingEventRegistration model
Previously we saved the `TeachingEventRegistration` as part of saving a
`Candidate`, but this was later changed so that the registration is saved
independently (a deep insert on `Candidate` doesn't work for reasons unknown).

The de-duplication logic was left in the `Candidate` model - this moves it into
the `TeachingEventRegistration` model, so if a registration already exists for
the user against the event a call to `ToEntity` returns `null` and the `Save`
operation on the `CrmService` just returns.