Refactor the Subscription models to match new schema
The subscriptions schema in the CRM has changed significantly; instead of being
a separate entity it has been flattened into the `contact` entity (which maps
to our `Candidate` model).

This commit removes the `Subscription` model, adds new subscription attributes
to the `Candidate` model and maps them to the CRM entity fields.