Return OnlineEvent type events in distance based searches
When a user searches for events that are a certain distance from a postcode we
currently do not return events that have the `OnlineEvent` type. This commit
updates the query to return these events for distance based searches. It
performs a second query to retrieve them; whilst it would be possible to do it
in a single query it would be much more verbose, and given the number of events
we are expecting there shouldn't be a noticable performance hit for doing two
queries.