Expose ChannelId for MailingListAddMember request
We want to be able to specify the channel from the front-end for on-campus
events (via a QR code and unique URL containing the channel id). This updates
the `MailingListAddMember` model to allow the `ChannelId` to be specified and
uses it for the candidate/mailing list/event subscription channels.

Add validation to the `Candidate` model to ensure the channel ids are valid
according to the CRM.