Add legacy support for HasTeacherTrainingAdviser check
If a candidate has already been assigned a teacher training adviser as part of
the current/legacy system we don't want them to be able to sign up for a TTA
again. We determine this by ensuring that their `contact.owningbusinessunit` is
not `1A61F629-F502-E911-A972-000D3A23443B` (a TTA).