Disable ETag caching on Teaching Event search
There is odd behaviour in the hosted environments around search results not
always returning correctly. The only difference I can think of with the
development environment is that caching is disabled (although I've tested with
it enabled in dev and it does seem to behave).

Disable caching in hosted environments to see if that clears up the unexpected
results; if it does we can investigate further and then re-enable caching.