Accept X-Client-IP header on all requests
Update the Swagger configuration to accept an `X-Client-IP` header on all
requests. The auto-generated library will allow this to be specified per-API
request, for example:

```
result = api_instance.get_callback_booking_quotas(x_client_ip: "1.2.3.4")
```

Going forward this IP address will be used to rate limit requests to the API,
in addition to the API key of the client.

If unspecified the requests will fall into the global client rate limiting
bucket.

Can be set to the end user IP address to allow per-user rate limiting in the
API. This is how we will be rate limiting individual users when they request a
PIN code email (so that they cannot abuse the GOV.UK Notify service to send
spam).