Add endpoint to get upcoming events by type
When a user lands on the events index page in the website we want to be able to
show them the next 3 events that are upcoming for each type. This improves the
user experience when we are approaching the end of the month and there are no
longer any events within the current month to display.

Adds a new endpoint `/api/teaching_events/upcoming_indexed_by_type` that will return
the upcoming teaching events grouped by their type.

The return type is `Dictionary<string, IEnumerable<TeachingEvent>>` due to
`System.Text.Json` not yet supporting non-string Dictionary keys (it would be
preferrable to use an `int` here).

Adds a `/api/teaching_events/search_indexed_by_type` for consistency, with a
view to removing the current `/search` endpoint once deprecated in the front
end.