Run EF Core migrations from CLI
EF Core advise against running the migrations at runtime in case two instances
attempt to execute them at the same time. They recommend exporting the
migrations to SQL and running them manually, which is overkill for what we need
so a happy medium is to run the via the CLI before the instance starts.