Run against Postgres DB in development/test
Updates the development and test environments to execute against a dockerized
Postgres database instance instead of using the in-memory database.

The advantages to do this are:

- Migrations are ran in the test suite/development.
- We can load and persist all the geolocation data in development.
- Development/test more closely match production.
- It's easier to run the app on a non-Windows machine.

The test environment works by creating a new database for each test that
requires one; we use the `gis_test` database that is created for us as a
template (after migrations have been ran).

Whilst it would be quicker to wrap each test run in a transaction using the
same database, this setup would be more complicated and using Postgres
templates seems to be fast enough for now.